+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 druid CHANGELOG
 ===============
 
+0.0.3
+-----
+* Updated default Druid version to 0.7.0
+* Moved archive filename spec to attributes
+* Now use new common config file where appropriate
+
 0.0.2
 -----
 * Replaced symbolic links with regular files